Create new razor homepage for logged out users (SF-481) - Initial landing page only includes first 2 sections and the contact form - Refactored home page to run from index - Setup Sass for razor pages - recommended with VC `Live Sass Compiler`
