--- src/SIL.XForge.Scripture/Pages/Home.cshtml
@@ -1,81 +1,0 @@
-@page
-@model HomepageModel
-@section css {
-    <link rel="stylesheet" href="~/css/sf.css" />
-    <style>
-        .mdl-grid {
-            padding:0px;
-        }
-    </style>
-}
-@section scripts {
-    <script type="text/javascript">
-        function logIn() {
-            var options = { state: JSON.stringify({}) };
-            webAuth.authorize(options);
-        }
-
-        function signUp() {
-            var options = { state: JSON.stringify({}), login_hint: 'signUp' };
-            webAuth.authorize(options);
-        }
-    </script>
-}
-
-<div class="mdl-layout__content">
-
-  <div id="header" class="flex-row prim-between sec-center mdl-button--raised">
-    <div id="title">
-      <img id="icon" src="~/sf.ico">
-      <span>Scripture Forge</span>
-    </div>
-    <div>
-    <a class="mdl-button mdl-button--raised mdl-button--colored mdl-button--accent" href="javascript:signUp()">
-          Sign up
-      </a>
-      <a class="mdl-button mdl-button--raised" href="javascript:logIn()">
-          Log in
-      </a>
-    </div>
-  </div>
-
-  <div id="content" class="flex-column prim-start sec-center"
-  ngClass.lt-md="smallcontent" ngClass.md="mediumcontent" ngClass.gt-md="largecontent">
-    <div id="welcome" class="flex-column prim-center sec-center">
-        <h1>Welcome to Scripture Forge</h1>
-        <p>Work with your Scripture data in a web app giving you the option of engaging a wider variety of language speakers. The online Scripture Checking App can help you run a Scripture checking campaign with the aim of getting broader feedback to the translation team quickly.</p>
-        <a class="mdl-button mdl-js-button mdl-button--raised mdl-button--colored" href="/register">Register your interest</a>
-    </div>
-    <div id="checking" class="flex-row prim-around sec-baseline">
-      <div id="offline" class="text-column">
-        <h4>Offline Capable</h4>
-        <p>Users can connect once to get the scripture content on their device and then answer questions offline. Their answers can then be uploaded when they have internet again.</p>
-      </div>
-      <div id="paratext" class="text-column">
-        <h4>Works with Paratext</h4>
-        <p>Send/Receive with Paratext to get the project into Scripture Forge using your Paratext credentials. Update Paratext with the notes generated by the community checking collaborators.</p>
-      </div>
-      <div id="engagement" class="text-column">
-        <h4>User Engagement</h4>
-        <p>Turn on "invite a friend" to open up your campaign or keep feedback to a defined group. Users can answer questions, comment on other's answers, and +1 other's responses. Generate user engagement reports at the end of your campaign.</p>
-      </div>
-    </div>
-    <div id="drafting">Pellentesque habitant morbi tristique senectus et netus. Amet purus gravida quis blandit turpis cursus in hac. Fames ac turpis egestas sed tempus urna et pharetra pharetra. Nisi vitae suscipit tellus mauris a diam maecenas sed. Faucibus pulvinar elementum integer enim neque. Odio tempor orci dapibus ultrices. Turpis tincidunt id aliquet risus feugiat in ante metus dictum. Tortor condimentum lacinia quis vel. Amet luctus venenatis lectus magna fringilla urna porttitor. Sit amet volutpat consequat mauris nunc congue nisi vitae. Posuere sollicitudin aliquam ultrices sagittis orci a scelerisque purus semper. Eu non diam phasellus vestibulum.</div>
-    <div id="footer" class="flex-column">
-        <div id="icons" class="flex-row prim-between sec-center">
-            <div id="payap"><a href="https://inter.payap.ac.th/"><img src="/images/logo-payap.svg" alt="Payap University Logo"></a></div>
-            <div id="sil"><a href="http://www.sil.org"><img src="/images/logo-sil.svg" alt="SIL International Logo"></a></div>
-        </div>
-        <div id="copyright" class="flex-row prim-center sec-center">
-            <div>
-              &copy; <span class="notranslate"> @DateTime.Now.Year </span><a target="_blank" href="http://www.sil.org">SIL International</a>
-              &nbsp;&nbsp;&nbsp;&nbsp;
-              v<span>@System.Diagnostics.FileVersionInfo.GetVersionInfo(@System.Reflection.Assembly.GetEntryAssembly().Location).ProductVersion</span>
-              &nbsp;&nbsp;&nbsp;&nbsp;
-              <a href="/terms_and_conditions">Terms</a>
-            </div>
-        </div>
-    </div>
-  </div>
-
-</div>
