+++ src/SIL.XForge.Scripture/Pages/Index.cshtml
@@ -1,17 +1,123 @@
 @page
 @model IndexModel
 @section scripts {
+    @await Html.PartialAsync("_ValidationScriptsPartial")
     <script type="text/javascript">
-        webAuth.checkSession({}, function(err, authResult) {
-            if (err) {
-                window.location.href = '/home';
-            } else {
-                var options = { state: JSON.stringify({}) };
-                webAuth.authorize(options);
-            }
-        });
+    document.getElementById("register-interest").addEventListener("click", function() {
+        window.location = "#contact";
+    });
+    webAuth.checkSession({}, function(err, authResult) {
+        if (!err) {
+            var options = { state: JSON.stringify({}) };
+            webAuth.authorize(options);
+        }
+    });
     </script>
 }
-<div style="margin: 50px 8px; font-family: sans-serif; line-height: 1.5; text-align:center; width: 100%">
-    Loading...
-</div>
+
+<section id="hero">
+    <div class="container">
+        <div class="mdl-grid">
+            <div class="mdl-cell mdl-cell--12-col align-center">
+                <img src="~/images/multi-devices.svg" alt="Works on multiple devices" />
+                <h1 class="align-center">A free set of tools to do more with your Paratext project, wherever you are.</h1>
+                <button id="register-interest" class="button">
+                    <span class="button-prefix">
+                        <span class="beta">Beta</span>
+                        <img src="~/images/logo-sf-white.svg" alt="Scripture Forge">
+                    </span>
+                    <span class="text">Register Your Interest</span>
+                </button>
+            </div>
+        </div>
+    </div>
+</section>
+<section id="engage">
+    <div class="container">
+        <div class="mdl-grid">
+            <div class="mdl-cell mdl-cell--12-col flex direction-column justify-content-center">
+                <div class="title align-center">Engage with the language community. Collaborate with your peers.</div>
+                    <div class="engage-features">
+                        <div class="engage-feature">
+                            <img src="~/images/logo-paratext.svg" alt="Paratext">
+                            <span>Collaborate on Scripture</span>
+                        </div>
+                        <div class="engage-feature">
+                            <img src="~/images/community-checking.svg" alt="Community Checking">
+                            <span>Do community checking</span>
+                        </div>
+                        <div class="engage-feature">
+                            <img src="~/images/translate.svg" alt="Translate">
+                            <span>Get suggestions as you type</span>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+</section>
+<section id="contact">
+    <div class="container">
+        <div class="mdl-grid">
+            <div class="mdl-cell mdl-cell--12-col">
+                <div class="section-title">
+                    <h2>Want to know more?</h2>
+                    <div class="title">Get in Touch With Us</div>
+                </div>
+                <div id="register" class="align-center">
+                    <h3>When can I start using Scripture Forge?</h3>
+                    <p>Send us your information and we will be in contact with you on how to join the beta.</p>
+                    <div asp-validation-summary="All" class="required"></div>
+                    <form method="POST">
+                        <div class="field-container">
+                            <div class="mdl-textfield mdl-js-textfield mdl-textfield--floating-label">
+                                <input class="mdl-textfield__input" asp-for="Name">
+                                <label class="mdl-textfield__label" for="Name">
+                                    What's your name and last name? <span class="required">*</span>
+                                </label>
+                                <span asp-validation-for="Name"></span>
+                            </div>
+                        </div>
+                        <div class="field-container">
+                            <div class="mdl-textfield mdl-js-textfield mdl-textfield--floating-label">
+                                <input class="mdl-textfield__input" asp-for="Email">
+                                <label class="mdl-textfield__label" for="Email">
+                                    Where can we email you? <span class="required">*</span>
+                                </label>
+                                <span asp-validation-for="Email"></span>
+                            </div>
+                        </div>
+                        <div class="field-container">
+                            <div class="mdl-textfield mdl-js-textfield mdl-textfield--floating-label">
+                                <select class="mdl-textfield__input" asp-for="Role">
+                                    <option></option>
+                                    <option value="Language Program Manager">Language Program Manager</option>
+                                    <option value="Translator">Translator</option>
+                                    <option value="Paratext Consultant">Paratext Consultant</option>
+                                    <option value="Language Technology Consultant">Language Technology Consultant</option>
+                                    <option value="Other">Other</option>
+                                </select>
+                                <label class="mdl-textfield__label" for="Role">What is your language project role?</label>
+                                <span asp-validation-for="Role"></span>
+                            </div>
+                        </div>
+                        <div class="field-container">
+                            <div class="mdl-textfield mdl-js-textfield mdl-textfield--floating-label">
+                                <textarea class="mdl-textfield__input" rows= "3" asp-for="Message"></textarea>
+                                <label class="mdl-textfield__label" for="Message">Why do you want to use Scripture Forge?</label>
+                                <span asp-validation-for="Message"></span>
+                            </div>
+                        </div>
+                        <button class="button" type="submit">
+                            <span class="button-prefix">
+                                <span class="beta">Beta</span>
+                                <img src="~/images/logo-sf-white.svg" alt="Scripture Forge">
+                            </span>
+                            <span class="text">Register</span>
+                        </button>
+                    </form>
+                </div>
+            </div>
+        </div>
+    </div>
+</section>
