+++ src/SIL.XForge.Scripture/wwwroot/scss/_header.scss
@@ -1,0 +1,22 @@
+header {
+  position: absolute;
+  width: 100%;
+  padding-top: 8px;
+
+  img {
+    width: 50px;
+    height: auto;
+    margin-right: 10px;
+  }
+
+  span {
+    font-size: 25px;
+  }
+
+  a {
+    color: $greyDark;
+    text-decoration: none;
+    display: flex;
+    align-items: center;
+  }
+}
