+++ src/SIL.XForge.Scripture/wwwroot/scss/_styles.scss
@@ -1,0 +1,106 @@
+body,
+p {
+  color: $greyDark;
+  font-weight: 300;
+  font-family: $fontFamily;
+  font-size: 15px;
+}
+h1,
+h2,
+h3,
+h4,
+h5,
+h6 {
+  font-weight: 300;
+}
+h1 {
+  font-size: 50px;
+  @include media-breakpoint-down(sm) {
+    font-size: 32px;
+  }
+}
+h2 {
+  font-size: 25px;
+  @include media-breakpoint-down(sm) {
+    font-size: 20px;
+  }
+}
+h3 {
+  font-size: 34px;
+  line-height: 1.2;
+  @include media-breakpoint-down(sm) {
+    font-size: 24px;
+  }
+}
+.title {
+  font-size: 40px;
+  line-height: 1.1;
+  @include media-breakpoint-down(sm) {
+    font-size: 30px;
+  }
+}
+
+img {
+  max-width: 100%;
+}
+
+.button {
+  background: $purpleDark;
+  font-size: 18px;
+  color: $white;
+  text-transform: uppercase;
+  font-family: $fontFamily;
+  border: 0;
+  cursor: pointer;
+  display: inline-flex;
+  padding: 0;
+  min-width: 280px;
+  text-decoration: none;
+  transition: background 0.25s;
+
+  &:hover {
+    background: darken($purpleDark, 10%);
+  }
+
+  .text {
+    padding: 26px;
+    flex: 1;
+    @include media-breakpoint-down(sm) {
+      padding: 16px;
+    }
+  }
+
+  .button-prefix {
+    background: $purpleMedium;
+    position: relative;
+    overflow: hidden;
+    display: flex;
+    align-self: normal;
+    align-items: center;
+
+    .beta {
+      text-transform: uppercase;
+      background: $purpleLight;
+      position: absolute;
+      top: 2px;
+      left: -28px;
+      transform: rotate(-43deg);
+      padding: 5px 30px;
+      font-size: 10px;
+      line-height: 1.3;
+      @include media-breakpoint-down(sm) {
+        top: 4px;
+        padding: 2px 30px;
+      }
+    }
+
+    img {
+      width: 26px;
+      padding: 0 26px;
+      height: 32px;
+      @include media-breakpoint-down(sm) {
+        padding: 0 16px;
+      }
+    }
+  }
+}
