+++ src/SIL.XForge.Scripture/wwwroot/scss/_variables.scss
@@ -1,0 +1,19 @@
+$greyLight: #a3a3a3;
+$greyDark: #343a31;
+$purpleMedium: #5a465f;
+$purpleLight: #a065e3;
+$purpleDark: #47314e;
+$greenLight: #b8d432;
+$greenDark: #5f7333;
+$white: #fff;
+$error: #cc0000;
+
+$fontFamily: Roboto, Arial, Helvetica, sans-serif;
+
+$grid-breakpoints: (
+  xs: 0,
+  sm: 576px,
+  md: 768px,
+  lg: 960px,
+  xl: 1140px
+);
